+++ lib/src/lottie_builder.dart
@@ -60,6 +60,7 @@ class LottieBuilder extends StatefulWidget {
     this.addRepaintBoundary,
     this.filterQuality,
     this.onWarning,
+    this.enableRenderCache,
   });
 
   /// Creates a widget that displays an [LottieComposition] obtained from the network.
@@ -86,6 +87,7 @@ class LottieBuilder extends StatefulWidget {
     this.filterQuality,
     this.onWarning,
     LottieDecoder? decoder,
+    this.enableRenderCache,
   }) : lottie = NetworkLottie(src,
             headers: headers,
             imageProviderFactory: imageProviderFactory,
@@ -123,6 +125,7 @@ class LottieBuilder extends StatefulWidget {
     this.filterQuality,
     this.onWarning,
     LottieDecoder? decoder,
+    this.enableRenderCache,
   }) : lottie = FileLottie(file,
             imageProviderFactory: imageProviderFactory, decoder: decoder);
 
@@ -151,6 +154,7 @@ class LottieBuilder extends StatefulWidget {
     this.filterQuality,
     this.onWarning,
     LottieDecoder? decoder,
+    this.enableRenderCache,
   }) : lottie = AssetLottie(name,
             bundle: bundle,
             package: package,
@@ -180,6 +184,7 @@ class LottieBuilder extends StatefulWidget {
     this.filterQuality,
     this.onWarning,
     LottieDecoder? decoder,
+    this.enableRenderCache,
   }) : lottie = MemoryLottie(bytes,
             imageProviderFactory: imageProviderFactory, decoder: decoder);
 
@@ -412,6 +417,27 @@ class LottieBuilder extends StatefulWidget {
   /// ```
   final ImageErrorWidgetBuilder? errorBuilder;
 
+  /// Opt-in a special render mode where the frames of the animation are
+  /// lazily rendered in offscreen images.
+  /// Subsequent runs of the animation will be very cheap to render.
+  ///
+  /// This is useful is the animation is complex and can consume lot of energy
+  /// from the battery.
+  /// This is will trade an excessive CPU usage for an increase memory usage.
+  ///
+  /// The render cache is managed internally and will release the memory once the
+  /// animation is disposed. The cache is shared between all animations. If 2 `Lottie`
+  /// widget are rendered at the same size, they will render only once.
+  ///
+  /// Any change in the configuration of the animation (delegates, frame rate etc...)
+  /// will clear the cache.
+  /// Any change in the size will invalidate the cache. The cache use the final size
+  /// visible on the screen (with all transforms applied).
+  ///
+  /// In order to not exceed the memory limit of a device, the cache is constrained
+  /// to maximum 100MiB. After that, animations are not cached anymore.
+  final bool? enableRenderCache;
+
   @override
   State<LottieBuilder> createState() => _LottieBuilderState();
 
@@ -507,6 +533,7 @@ class _LottieBuilderState extends State<LottieBuilder> {
           alignment: widget.alignment,
           addRepaintBoundary: widget.addRepaintBoundary,
           filterQuality: widget.filterQuality,
+          enableRenderCache: widget.enableRenderCache,
         );
 
         if (widget.frameBuilder != null) {